handle merging issues

+++ core/domain/src/main/java/org/the_chance/honeymart/domain/model/CartEntity.kt
@@ -8,6 +8,6 @@ data class CartEntity(
 data class CartProductsEntity(
     val id: Long?,
     val name: String?,
-    val count: Long?,
+    val count: Int?,
     val price: Double?,
 )
